migrate: move util (HTMLComment) to components directory
- Migrated util component files to src/components/util/
- Moved and renamed HTMLComment.tsx -> template.tsx
- Moved and renamed HTMLComment.test.tsx -> template.test.tsx
- Updated import statements in:
  - src/pages/[[...slug]].tsx from @/templates/common/util/HTMLComment to @/components/util/template
  - src/components/footer/template.tsx from @/templates/common/util/HTMLComment to @/components/util/template
- Fixed internal import in template.test.tsx from ./HTMLComment to ./template
- Removed empty src/templates/common/util/ directory
- All TypeScript compilation checks passing
- All tests passing (3 tests for HTMLComment component)

This util component contains the HTMLComment utility that dynamically injects HTML comments
into the document head or body for SEO and debugging purposes. It's used by footer and
main page layouts to add VA mission statement comments.
